Specify Normalized Paths
Also clarify the relationship of JSONPath in general, and Normalized
Paths in particular, to JSON Pointer.

The approach for Normal Paths is to minimise escaping. The alternative
of maximising escaping was rejected because the results are unreadable.
The alternative of escaping all but a particular set of characters was
rejected because it would be biassed.

Co-authored-by: cabo <cabo@tzi.org>

Fixes https://github.com/ietf-wg-jsonpath/draft-ietf-jsonpath-base/issues/67