Delete confusing language about Normalized Paths
The one to one correspondence sentence gives us most of the desired
properties. The fact that for distinct Normalized Paths there will always
be at least one JSON value which can distinguish the paths is interesting
but not crucial. If it's confusing to some people, then it's probably
a distraction.