Simplify well-typedness condition
When the declared type of the parameter is LogicalType
and the argument is a logical-expr there are just two
cases to deal with. Either the argument is a function
expression and it must have a result of type NodesType
or (argument type same as parameter type) LogicalType
or the argument is a logical-expr that is not a function
expression.

However, the well-typeness of a logical-expr that is not
a function expression is already covered by (1) in 2.4.3.