Rob Wilton #1: clarify luxury overflow
The specific case addressed here is about JSON values large enough so
you no longer can evaluate [?length(@) < length($.a)].
This is a bit of a luxury problem, as 2**53 bytes ⩰ 9 Petabytes.
Backblaze reports they can [store a Petabyte for some $35000][1], so
this is not an entirely hypothetical case any more, and will get less
so during the time the spec lasts.

[1]: https://www.backblaze.com/blog/petabytes-on-a-budget-10-years-and-counting/